+++ forui_hooks/lib/src/popover_controller_hook.dart
@@ -8,13 +8,11 @@ import 'package:forui/forui.dart';
 FPopoverController useFPopoverController({
   TickerProvider? vsync,
   bool shown = false,
-  FPopoverMotion motion = const FPopoverMotion(),
   List<Object?>? keys,
 }) => use(
   _PopoverControllerHook(
     vsync: vsync ??= useSingleTickerProvider(keys: keys),
     shown: shown,
-    motion: motion,
     keys: keys,
   ),
 );
@@ -22,9 +20,8 @@ FPopoverController useFPopoverController({
 class _PopoverControllerHook extends Hook<FPopoverController> {
   final TickerProvider vsync;
   final bool shown;
-  final FPopoverMotion motion;
 
-  const _PopoverControllerHook({required this.vsync, required this.shown, required this.motion, super.keys});
+  const _PopoverControllerHook({required this.vsync, required this.shown, super.keys});
 
   @override
   _PopoverControllerHookState createState() => .new();
@@ -34,13 +31,12 @@ class _PopoverControllerHook extends Hook<FPopoverController> {
     super.debugFillProperties(properties);
     properties
       ..add(DiagnosticsProperty('vsync', vsync))
-      ..add(FlagProperty('shown', value: shown, ifTrue: 'shown'))
-      ..add(DiagnosticsProperty('motion', motion));
+      ..add(FlagProperty('shown', value: shown, ifTrue: 'shown'));
   }
 }
 
 class _PopoverControllerHookState extends HookState<FPopoverController, _PopoverControllerHook> {
-  late final _controller = FPopoverController(vsync: hook.vsync, shown: hook.shown, motion: hook.motion);
+  late final _controller = FPopoverController(vsync: hook.vsync, shown: hook.shown);
 
   @override
   FPopoverController build(BuildContext context) => _controller;